+++ src/torchmetrics/functional/segmentation/dice.py
@@ -1,4 +1,4 @@
-# Copyright The Lightning team.
+# Copyright The PyTorch Lightning team.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,8 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from torchmetrics.segmentation.dice import DiceScore
-from torchmetrics.segmentation.generalized_dice import GeneralizedDiceScore
-from torchmetrics.segmentation.mean_iou import MeanIoU
+import torch
+from torch import Tensor
+from typing_extensions import Literal
 
-__all__ = ["GeneralizedDiceScore", "MeanIoU", "DiceScore"]
+from torchmetrics.functional.segmentation.utils import _ignore_background
+from torchmetrics.utilities.checks import _check_same_shape
+from torchmetrics.utilities.compute import _safe_divide
+
+
+def dice_score():
+    pass